RavenDB-17793 : fixes after rebase
